ISSD-1352 Refactoring Product Card
