+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/components/RadioCardList/RadioCardList.tsx
@@ -37,7 +37,7 @@ const RadioCardList = <T extends unknown>({
 
 	return (
 		<div className="d-flex flex-column w-100">
-			<div className="mb-0 pr-3 w-100">
+			<div className="mb-0 w-100">
 				{contentList.map((content, index) => (
 					<RadioCard
 						activeRadio={content.selected}
